added the posted-by tab and the picture styling for the project page
